requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
dnspython==2.2.1
email-validator==1.2.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
